requirements.txt
# One-Agent Dependencies
# Core SDK
claude-agent-sdk==0.1.29

# LLM Providers
anthropic>=0.25.0
openai>=1.0.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic>=2.0.0

# Utilities
requests>=2.31.0
colorama>=0.4.6

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
